# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /site/seed-guide.pdf:
    get:
      operationId: downloadSeedGuidePdf
      tags: [catalogue]
      summary: Download the uploaded pasture seed guide PDF
      responses:
        "200":
          description: Seed guide PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { description: No seed guide PDF has been uploaded }
  /admin/site-settings:
    get:
      operationId: getAdminSiteSettings
      tags: [admin]
      summary: Get homepage, about page, company and seed guide settings for editing
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Current site settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteSettings" }
        "401": { description: Sign in is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "403": { description: Administrator access is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    put:
      operationId: updateSiteSettings
      tags: [admin]
      summary: Save homepage, about page, company and seed guide settings
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SiteSettingsInput" }
      responses:
        "200":
          description: Saved site settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteSettings" }
        "400": { description: Invalid site settings, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: Sign in is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "403": { description: Administrator access is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /admin/site-settings/seed-guide-pdf:
    post:
      operationId: uploadSeedGuidePdf
      tags: [admin]
      summary: Upload or replace the public pasture seed guide PDF
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SeedGuidePdfInput" }
      responses:
        "200":
          description: Site settings including the new PDF
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteSettings" }
        "400": { description: Invalid PDF, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: Sign in is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "403": { description: Administrator access is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    SiteHeroImage:
      type: object
      required: [src, assetId]
      properties:
        src: { type: string, maxLength: 500 }
        assetId: { type: string, nullable: true, maxLength: 80 }
    SiteHomepageSettings:
      type: object
      required: [heroImageSrc, heroImageAssetId, heroImages, heroSlideshow, heroEyebrow, heroHeading, heroBody, aboutBody, bestSellerSlugs]
      properties:
        heroImageSrc: { type: string, maxLength: 500 }
        heroImageAssetId: { type: string, nullable: true, maxLength: 80 }
        heroImages:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/SiteHeroImage" }
        heroSlideshow: { type: boolean }
        heroEyebrow: { type: string, maxLength: 120 }
        heroHeading: { type: string, maxLength: 180 }
        heroBody: { type: string, maxLength: 2000 }
        aboutBody: { type: string, maxLength: 2000 }
        bestSellerSlugs:
          type: array
          maxItems: 4
          items: { type: string, maxLength: 160 }
    SiteSeedGuideSettings:
      type: object
      required: [navTitle, cardHeading, cardButtonLabel, cardImageSrc, cardImageAssetId, pdfFilename, pdfPublicUrl, pageTitle, pageIntro, pageButtonLabel]
      properties:
        navTitle: { type: string, maxLength: 80 }
        cardHeading: { type: string, maxLength: 240 }
        cardButtonLabel: { type: string, maxLength: 120 }
        cardImageSrc: { type: string, maxLength: 500 }
        cardImageAssetId: { type: string, nullable: true, maxLength: 80 }
        pdfFilename: { type: string, maxLength: 160 }
        pdfPublicUrl: { type: string, maxLength: 500 }
        pageTitle: { type: string, maxLength: 180 }
        pageIntro: { type: string, maxLength: 2000 }
        pageButtonLabel: { type: string, maxLength: 120 }
    SiteAboutValue:
      type: object
      required: [title, body]
      properties:
        title: { type: string, maxLength: 80 }
        body: { type: string, maxLength: 500 }
    SiteAboutSettings:
      type: object
      required: [heroEyebrow, heroHeading, heroHeadingEmphasis, heroIntro, heroImageSrc, heroImageAssetId, storyLead, storyParagraphs, valuesHeading, valuesHeadingEmphasis, values]
      properties:
        heroEyebrow: { type: string, maxLength: 120 }
        heroHeading: { type: string, maxLength: 180 }
        heroHeadingEmphasis: { type: string, maxLength: 80 }
        heroIntro: { type: string, maxLength: 500 }
        heroImageSrc: { type: string, maxLength: 500 }
        heroImageAssetId: { type: string, nullable: true, maxLength: 80 }
        storyLead: { type: string, maxLength: 2000 }
        storyParagraphs:
          type: array
          maxItems: 3
          items: { type: string, maxLength: 4000 }
        valuesHeading: { type: string, maxLength: 80 }
        valuesHeadingEmphasis: { type: string, maxLength: 80 }
        values:
          type: array
          maxItems: 3
          items: { $ref: "#/components/schemas/SiteAboutValue" }
    SiteCompanySettings:
      type: object
      required: [legalName, tradingName, phone, email, address, officeHours, abn]
      properties:
        legalName: { type: string, maxLength: 160 }
        tradingName: { type: string, maxLength: 160 }
        phone: { type: string, maxLength: 40 }
        email: { type: string, maxLength: 180 }
        address: { type: string, maxLength: 240 }
        officeHours: { type: string, maxLength: 120 }
        abn: { type: string, maxLength: 20 }
    SiteSettings:
      type: object
      required: [homepage, seedGuide, about, company, updatedAt]
      properties:
        homepage: { $ref: "#/components/schemas/SiteHomepageSettings" }
        seedGuide: { $ref: "#/components/schemas/SiteSeedGuideSettings" }
        about: { $ref: "#/components/schemas/SiteAboutSettings" }
        company: { $ref: "#/components/schemas/SiteCompanySettings" }
        updatedAt: { type: string, format: date-time }
    SiteSeedGuideInput:
      type: object
      required: [navTitle, cardHeading, cardButtonLabel, cardImageSrc, cardImageAssetId, pageTitle, pageIntro, pageButtonLabel]
      properties:
        navTitle: { type: string, maxLength: 80 }
        cardHeading: { type: string, maxLength: 240 }
        cardButtonLabel: { type: string, maxLength: 120 }
        cardImageSrc: { type: string, maxLength: 500 }
        cardImageAssetId: { type: string, nullable: true, maxLength: 80 }
        pageTitle: { type: string, maxLength: 180 }
        pageIntro: { type: string, maxLength: 2000 }
        pageButtonLabel: { type: string, maxLength: 120 }
    SiteSettingsInput:
      type: object
      required: [homepage, seedGuide]
      properties:
        homepage: { $ref: "#/components/schemas/SiteHomepageSettings" }
        seedGuide: { $ref: "#/components/schemas/SiteSeedGuideInput" }
        about: { $ref: "#/components/schemas/SiteAboutSettings" }
        company: { $ref: "#/components/schemas/SiteCompanySettings" }
    SeedGuidePdfInput:
      type: object
      required: [filename, data]
      properties:
        filename: { type: string, minLength: 1, maxLength: 160 }
        data: { type: string, minLength: 1 }
...