# lib/api-zod/src/generated/types/siteAboutValue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteAboutValue {
  /** @maxLength 80 */
  title: string;
  /** @maxLength 500 */
  body: string;
}