# lib/api-zod/src/generated/types/siteCompanySettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteCompanySettings {
  /** @maxLength 160 */
  legalName: string;
  /** @maxLength 160 */
  tradingName: string;
  /** @maxLength 40 */
  phone: string;
  /** @maxLength 180 */
  email: string;
  /** @maxLength 240 */
  address: string;
  /** @maxLength 120 */
  officeHours: string;
  /** @maxLength 20 */
  abn: string;
}