# lib/api-zod/src/generated/types/siteSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SiteAboutSettings } from './siteAboutSettings';
import type { SiteCompanySettings } from './siteCompanySettings';
import type { SiteHomepageSettings } from './siteHomepageSettings';
import type { SiteSeedGuideInput } from './siteSeedGuideInput';

export interface SiteSettingsInput {
  homepage: SiteHomepageSettings;
  seedGuide: SiteSeedGuideInput;
  about?: SiteAboutSettings;
  company?: SiteCompanySettings;
}